MFLP-26: ajuste nas pastas
