--- src/routes/backoffice/usuario.ts
@@ -1,8 +1,0 @@
-import { Router } from "express";
-import { listarUsuariosController } from "../../controllers/backoffice/listarUsuarios";
-
-const router = Router();
-
-router.post("/backoffice/usuarios", listarUsuariosController);
-
-export default router;
